==> nbs/15_gsa_gls/99-gls-export-phenoplier.ipynb <==
# Description

This notebook exports results into other more accessible data formats.

# Modules

from pathlib import Path
import shutil

import pandas as pd
import rpy2.robjects as ro
from rpy2.robjects import pandas2ri
from rpy2.robjects.conversion import localconverter

import conf
from utils import remove_all_file_extensions

readRDS = ro.r["readRDS"]

saveRDS = ro.r["saveRDS"]

# Settings

OUTPUT_DIR = Path(conf.__file__).parent.parent / "data" / "gls"
display(OUTPUT_DIR)
OUTPUT_DIR.mkdir(exist_ok=True, parents=True)

# Get results files

input_filepath = conf.RESULTS["GLS"] / "gls-summary-phenomexcan.pkl.gz"
display(input_filepath)

data = pd.read_pickle(input_filepath)

data.shape

data.head()

assert not data.isna().any().any()

# Save

## Pickle format

shutil.copy(input_filepath, OUTPUT_DIR)

## RDS format

output_file = remove_all_file_extensions(input_filepath).with_suffix(".rds")
output_file = OUTPUT_DIR / output_file.name
display(output_file)

with localconverter(ro.default_converter + pandas2ri.converter):
    data_r = ro.conversion.py2rpy(data)

data_r

saveRDS(data_r, str(output_file))

# testing: load the rds file again
data_r = readRDS(str(output_file))

with localconverter(ro.default_converter + pandas2ri.converter):
    data_again = ro.conversion.rpy2py(data_r)
    data_again.index = data_again.index.astype(int)

data_again.shape

data_again.head()

pd.testing.assert_frame_equal(
    data,
    data_again,
    check_dtype=False,
)

## Text format

# tsv format
output_file = remove_all_file_extensions(input_filepath).with_suffix(".tsv.gz")
output_file = OUTPUT_DIR / output_file.name
display(output_file)

data.to_csv(output_file, sep="\t", index=False, float_format="%.5e")

# testing
data2 = data.copy()
data2.index = list(range(0, data2.shape[0]))

data_again = pd.read_csv(output_file, sep="\t")
data_again.index = list(data_again.index)

data_again.shape

data_again.head()

pd.testing.assert_frame_equal(
    data2,
    data_again,
    check_categorical=False,
    check_dtype=False,
)
